services/auth-services/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Pydantic (FastAPI dependency)
pydantic==2.7.1
pydantic-settings==2.2.1

# Optional: Alembic if you need migrations in auth-service
alembic==1.13.1
